typedb 2.29.1
The `typedb` formula hasn't been updated since 2.23.0, as upstream
now splits release assets into arm64 and x86_64 archives (so `brew
bump-formula-pr` produces a broken `url`). This updates to the latest
version, 2.29.1, but it requires a somewhat ridiculous approach (the
cask DSL has a more elegant way of handling some of this).

Besides that, the license changed to MPL 2.0 in version 2.28.0.